# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Portfolio API with content management
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export interface LoginResponse {
  token: string;
  message: string;
}

export interface AuthStatus {
  authenticated: boolean;
  /** @nullable */
  username: string | null;
}

export interface HeroContent {
  name: string;
  title: string;
  tagline: string;
  email: string;
  phone: string;
  linkedin: string;
  github: string;
  availableForWork: boolean;
}

export interface ExperienceItem {
  id: string;
  company: string;
  /** @nullable */
  payroll: string | null;
  role: string;
  startDate: string;
  /** @nullable */
  endDate: string | null;
  location: string;
  bullets: string[];
}

export interface ExperienceContent {
  items: ExperienceItem[];
}

export interface ProjectItem {
  id: string;
  title: string;
  techStack: string;
  description: string;
  bullets: string[];
  /** @nullable */
  githubUrl: string | null;
  /** @nullable */
  period: string | null;
}

export interface ProjectsContent {
  items: ProjectItem[];
}

export interface SkillCategory {
  id: string;
  title: string;
  skills: string[];
}

export interface SkillsContent {
  categories: SkillCategory[];
}

export interface EducationItem {
  id: string;
  institution: string;
  degree: string;
  period: string;
  location: string;
  /** @nullable */
  percentage: string | null;
}

export interface EducationContent {
  items: EducationItem[];
}

export interface AchievementItem {
  id: string;
  title: string;
  description: string;
}

export interface AchievementsContent {
  items: AchievementItem[];
}

export interface PortfolioContent {
  hero: HeroContent;
  experience: ExperienceContent;
  projects: ProjectsContent;
  skills: SkillsContent;
  education: EducationContent;
  achievements: AchievementsContent;
}