# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Portfolio API with content management
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Admin login
 */
export const LoginBody = zod.object({
  username: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  token: zod.string(),
  message: zod.string(),
});

/**
 * @summary Logout
 */
export const LogoutResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Get auth status
 */
export const GetMeResponse = zod.object({
  authenticated: zod.boolean(),
  username: zod.string().nullable(),
});

/**
 * @summary Get all portfolio content
 */
export const GetPortfolioContentResponse = zod.object({
  hero: zod.object({
    name: zod.string(),
    title: zod.string(),
    tagline: zod.string(),
    email: zod.string(),
    phone: zod.string(),
    linkedin: zod.string(),
    github: zod.string(),
    availableForWork: zod.boolean(),
  }),
  experience: zod.object({
    items: zod.array(
      zod.object({
        id: zod.string(),
        company: zod.string(),
        payroll: zod.string().nullable(),
        role: zod.string(),
        startDate: zod.string(),
        endDate: zod.string().nullable(),
        location: zod.string(),
        bullets: zod.array(zod.string()),
      }),
    ),
  }),
  projects: zod.object({
    items: zod.array(
      zod.object({
        id: zod.string(),
        title: zod.string(),
        techStack: zod.string(),
        description: zod.string(),
        bullets: zod.array(zod.string()),
        githubUrl: zod.string().nullable(),
        period: zod.string().nullable(),
      }),
    ),
  }),
  skills: zod.object({
    categories: zod.array(
      zod.object({
        id: zod.string(),
        title: zod.string(),
        skills: zod.array(zod.string()),
      }),
    ),
  }),
  education: zod.object({
    items: zod.array(
      zod.object({
        id: zod.string(),
        institution: zod.string(),
        degree: zod.string(),
        period: zod.string(),
        location: zod.string(),
        percentage: zod.string().nullable(),
      }),
    ),
  }),
  achievements: zod.object({
    items: zod.array(
      zod.object({
        id: zod.string(),
        title: zod.string(),
        description: zod.string(),
      }),
    ),
  }),
});

/**
 * @summary Update hero section
 */
export const UpdateHeroContentBody = zod.object({
  name: zod.string(),
  title: zod.string(),
  tagline: zod.string(),
  email: zod.string(),
  phone: zod.string(),
  linkedin: zod.string(),
  github: zod.string(),
  availableForWork: zod.boolean(),
});

export const UpdateHeroContentResponse = zod.object({
  name: zod.string(),
  title: zod.string(),
  tagline: zod.string(),
  email: zod.string(),
  phone: zod.string(),
  linkedin: zod.string(),
  github: zod.string(),
  availableForWork: zod.boolean(),
});

/**
 * @summary Update experience section
 */
export const UpdateExperienceContentBody = zod.object({
  items: zod.array(
    zod.object({
      id: zod.string(),
      company: zod.string(),
      payroll: zod.string().nullable(),
      role: zod.string(),
      startDate: zod.string(),
      endDate: zod.string().nullable(),
      location: zod.string(),
      bullets: zod.array(zod.string()),
    }),
  ),
});

export const UpdateExperienceContentResponse = zod.object({
  items: zod.array(
    zod.object({
      id: zod.string(),
      company: zod.string(),
      payroll: zod.string().nullable(),
      role: zod.string(),
      startDate: zod.string(),
      endDate: zod.string().nullable(),
      location: zod.string(),
      bullets: zod.array(zod.string()),
    }),
  ),
});

/**
 * @summary Update projects section
 */
export const UpdateProjectsContentBody = zod.object({
  items: zod.array(
    zod.object({
      id: zod.string(),
      title: zod.string(),
      techStack: zod.string(),
      description: zod.string(),
      bullets: zod.array(zod.string()),
      githubUrl: zod.string().nullable(),
      period: zod.string().nullable(),
    }),
  ),
});

export const UpdateProjectsContentResponse = zod.object({
  items: zod.array(
    zod.object({
      id: zod.string(),
      title: zod.string(),
      techStack: zod.string(),
      description: zod.string(),
      bullets: zod.array(zod.string()),
      githubUrl: zod.string().nullable(),
      period: zod.string().nullable(),
    }),
  ),
});

/**
 * @summary Update skills section
 */
export const UpdateSkillsContentBody = zod.object({
  categories: zod.array(
    zod.object({
      id: zod.string(),
      title: zod.string(),
      skills: zod.array(zod.string()),
    }),
  ),
});

export const UpdateSkillsContentResponse = zod.object({
  categories: zod.array(
    zod.object({
      id: zod.string(),
      title: zod.string(),
      skills: zod.array(zod.string()),
    }),
  ),
});

/**
 * @summary Update education section
 */
export const UpdateEducationContentBody = zod.object({
  items: zod.array(
    zod.object({
      id: zod.string(),
      institution: zod.string(),
      degree: zod.string(),
      period: zod.string(),
      location: zod.string(),
      percentage: zod.string().nullable(),
    }),
  ),
});

export const UpdateEducationContentResponse = zod.object({
  items: zod.array(
    zod.object({
      id: zod.string(),
      institution: zod.string(),
      degree: zod.string(),
      period: zod.string(),
      location: zod.string(),
      percentage: zod.string().nullable(),
    }),
  ),
});

/**
 * @summary Update achievements section
 */
export const UpdateAchievementsContentBody = zod.object({
  items: zod.array(
    zod.object({
      id: zod.string(),
      title: zod.string(),
      description: zod.string(),
    }),
  ),
});

export const UpdateAchievementsContentResponse = zod.object({
  items: zod.array(
    zod.object({
      id: zod.string(),
      title: zod.string(),
      description: zod.string(),
    }),
  ),
});