# lib/api-zod/src/generated/types/achievementItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Portfolio API with content management
 * OpenAPI spec version: 0.1.0
 */

export interface AchievementItem {
  id: string;
  title: string;
  description: string;
}