# lib/api-zod/src/generated/types/achievementsContent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Portfolio API with content management
 * OpenAPI spec version: 0.1.0
 */
import type { AchievementItem } from "./achievementItem";

export interface AchievementsContent {
  items: AchievementItem[];
}